Updated deps (see list below) and fixed some new coding standard and static analysis violations
  - Locking symfony/polyfill-php84 (v1.33.0)
  - Upgrading aws/aws-sdk-php (3.342.23 => 3.356.37)
  - Upgrading bitexpert/phpstan-magento (v0.41.0 => v0.42.0)
  - Upgrading brick/math (0.12.3 => 0.14.0)
  - Upgrading colinmollenhour/credis (v1.16.2 => v1.17.0)
  - Upgrading colinmollenhour/php-redis-session-abstract (v2.0.0 => v2.1.2)
  - Upgrading composer/ca-bundle (1.5.6 => 1.5.8)
  - Upgrading composer/class-map-generator (1.6.1 => 1.6.2)
  - Upgrading composer/composer (2.8.8 => 2.8.12)
  - Upgrading composer/semver (3.4.3 => 3.4.4)
  - Upgrading composer/spdx-licenses (1.5.8 => 1.5.9)
  - Upgrading dealerdirect/phpcodesniffer-composer-installer (v1.0.0 => v1.1.2)
  - Upgrading ergebnis/composer-normalize (2.45.0 => 2.48.2)
  - Upgrading ergebnis/json (1.4.0 => 1.6.0)
  - Upgrading ergebnis/json-normalizer (4.8.0 => 4.10.1)
  - Upgrading ergebnis/json-pointer (3.6.0 => 3.7.1)
  - Upgrading ergebnis/json-printer (3.7.0 => 3.8.1)
  - Upgrading ergebnis/json-schema-validator (4.4.0 => 4.5.1)
  - Upgrading fidry/cpu-core-counter (1.2.0 => 1.3.0)
  - Upgrading friendsofphp/php-cs-fixer (v3.75.0 => v3.88.2)
  - Upgrading guzzlehttp/guzzle (7.9.3 => 7.10.0)
  - Upgrading guzzlehttp/promises (2.2.0 => 2.3.0)
  - Upgrading guzzlehttp/psr7 (2.7.1 => 2.8.0)
  - Upgrading justinrainbow/json-schema (6.4.1 => 6.6.0)
  - Upgrading laminas/laminas-escaper (2.16.0 => 2.17.0)
  - Upgrading laminas/laminas-filter (2.40.0 => 2.41.0)
  - Upgrading laminas/laminas-http (2.21.0 => 2.22.0)
  - Upgrading laminas/laminas-i18n (2.29.0 => 2.30.0)
  - Upgrading laminas/laminas-servicemanager (3.23.0 => 3.23.1)
  - Upgrading laminas/laminas-session (2.24.0 => 2.25.1)
  - Upgrading laminas/laminas-validator (2.64.2 => 2.64.4)
  - Upgrading league/flysystem (3.29.1 => 3.30.0)
  - Upgrading league/flysystem-local (3.29.0 => 3.30.0)
  - Upgrading localheinz/diff (1.2.0 => 1.3.0)
  - Upgrading magento/framework (103.0.8 => 103.0.8-p2)
  - Upgrading magento/magento-coding-standard (36 => 38)
  - Upgrading magento/module-bundle (101.0.8 => 101.0.8-p2)
  - Upgrading magento/module-catalog (104.0.8 => 104.0.8-p1)
  - Upgrading magento/module-catalog-rule (101.2.8 => 101.2.8-p1)
  - Upgrading magento/module-cms (104.0.8 => 104.0.8-p1)
  - Upgrading magento/module-customer (103.0.8 => 103.0.8-p2)
  - Upgrading magento/module-downloadable (100.4.8 => 100.4.8-p2)
  - Upgrading magento/module-email (101.1.8 => 101.1.8-p2)
  - Upgrading magento/module-newsletter (100.4.8 => 100.4.8-p2)
  - Upgrading magento/module-quote (101.2.8 => 101.2.8-p1)
  - Upgrading magento/module-sales (103.0.8 => 103.0.8-p2)
  - Upgrading magento/module-sales-rule (101.2.8 => 101.2.8-p1)
  - Upgrading magento/module-shipping (100.4.8 => 100.4.8-p2)
  - Upgrading magento/module-store (101.1.8 => 101.1.8-p2)
  - Upgrading magento/module-theme (101.1.8 => 101.1.8-p1)
  - Upgrading magento/module-user (101.2.8 => 101.2.8-p2)
  - Upgrading magento/zend-pdf (1.16.4 => 1.16.5)
  - Upgrading marc-mabe/php-enum (v4.7.1 => v4.7.2)
  - Upgrading paragonie/constant_time_encoding (v3.0.0 => v3.1.3)
  - Upgrading phpcompatibility/php-compatibility (dev-develop 9013cd0 => dev-develop 8daeec5)
  - Upgrading phpcsstandards/phpcsutils (1.0.12 => 1.1.2)
  - Upgrading phpseclib/phpseclib (3.0.43 => 3.0.47)
  - Upgrading phpstan/phpstan (1.12.23 => 1.12.32)
  - Upgrading phpstan/phpstan (2.1.11 => 2.1.31)
  - Upgrading phpstan/phpstan-deprecation-rules (2.0.1 => 2.0.3)
  - Upgrading phpstan/phpstan-strict-rules (2.0.4 => 2.0.7)
  - Upgrading ramsey/uuid (4.7.6 => 4.9.1)
  - Upgrading react/promise (v3.2.0 => v3.3.0)
  - Upgrading squizlabs/php_codesniffer (3.12.1 => 3.13.4)
  - Upgrading symfony/console (v6.4.20 => v6.4.26)
  - Upgrading symfony/console (v7.2.5 => v7.3.4)
  - Upgrading symfony/deprecation-contracts (v3.5.1 => v3.6.0)
  - Upgrading symfony/event-dispatcher (v7.2.0 => v7.3.3)
  - Upgrading symfony/event-dispatcher-contracts (v3.5.1 => v3.6.0)
  - Upgrading symfony/filesystem (v7.2.0 => v7.3.2)
  - Upgrading symfony/finder (v7.2.2 => v7.3.2)
  - Upgrading symfony/intl (v6.4.15 => v6.4.26)
  - Upgrading symfony/mailer (v6.4.18 => v6.4.26)
  - Upgrading symfony/mime (v6.4.19 => v6.4.26)
  - Upgrading symfony/options-resolver (v7.2.0 => v7.3.3)
  - Upgrading symfony/polyfill-ctype (v1.31.0 => v1.33.0)
  - Upgrading symfony/polyfill-intl-grapheme (v1.31.0 => v1.33.0)
  - Upgrading symfony/polyfill-intl-idn (v1.31.0 => v1.33.0)
  - Upgrading symfony/polyfill-intl-normalizer (v1.31.0 => v1.33.0)
  - Upgrading symfony/polyfill-mbstring (v1.31.0 => v1.33.0)
  - Upgrading symfony/polyfill-php73 (v1.31.0 => v1.33.0)
  - Upgrading symfony/polyfill-php80 (v1.31.0 => v1.33.0)
  - Upgrading symfony/polyfill-php81 (v1.31.0 => v1.33.0)
  - Upgrading symfony/process (v6.4.20 => v6.4.26)
  - Upgrading symfony/process (v7.2.5 => v7.3.4)
  - Upgrading symfony/service-contracts (v3.5.1 => v3.6.0)
  - Upgrading symfony/stopwatch (v7.2.4 => v7.3.0)
  - Upgrading symfony/string (v7.2.0 => v7.3.4)
  - Upgrading tedivm/jshrink (v1.7.0 => v1.8.0)
  - Upgrading webonyx/graphql-php (v15.20.0 => v15.25.1)
  - Upgrading wikimedia/less.php (v5.2.2 => v5.4.0)

+++ Console/UserInteraction.php
@@ -188,7 +188,7 @@ class UserInteraction
 
         $bytes = max($bytes, 0);
         $pow = floor(($bytes > 0 ? log($bytes) : 0) / log(1024));
-        $pow = min($pow, count($units) - 1);
+        $pow = (int) min($pow, count($units) - 1);
 
         $bytes /= (1 << (10 * $pow));
